requirements.txt
# ============================================
# Triton Inference Server Client
# ============================================
tritonclient[http]==2.63.0

# ============================================
# Deep Learning Frameworks
# ============================================
torch==2.3.1
torchvision==0.18.1
transformers==4.30.0

# ============================================
# Image Processing
# ============================================
Pillow==10.0.0

# ============================================
# Data Science & Utilities
# ============================================
numpy>=1.24.0,<2.0.0
requests>=2.31.0

# ============================================
# Jupyter Notebooks (Development)
# ============================================
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# ============================================
# Load Testing
# ============================================
locust==2.15.1
